requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
Jinja2==3.1.4
playwright==1.44.0
nltk==3.8.1
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
lxml-html-clean==0.4.3
requests
beautifulsoup4
flask-limiter
